Implemented the Category exception for exception

+++ src/com/library/system/service/impl/CategoryServiceImpl.java
@@ -1,5 +1,7 @@
 package com.library.system.service.impl;
 
+import com.library.system.exception.categoryException.CategoryAlreadyExistsException;
+import com.library.system.exception.categoryException.CategoryNotFoundException;
 import com.library.system.model.Category;
 import com.library.system.repository.CategoryRepository;
 import com.library.system.repository.impl.CategoryRepositoryImpl;
@@ -20,10 +22,10 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
-    public void addCategory(Category category) throws SQLException {
-        // Ajouter une vérification pour l'existence de la catégorie
+    public void addCategory(Category category) throws SQLException, CategoryAlreadyExistsException {
+        // Vérification si la catégorie existe déjà
         if (categoryRepository.doesCategoryExist(category.getCategory_name())) {
-            System.out.println("La catégorie existe déjà : " + category.getCategory_name());
+            throw new CategoryAlreadyExistsException("La catégorie existe déjà : " + category.getCategory_name());
         } else {
             categoryRepository.addCategory(category);
         }
@@ -36,8 +38,10 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public boolean doesCategoryExist(String categoryName) throws SQLException {
+        // Normaliser la casse du nom de la catégorie en minuscules
+        String normalizedCategoryName = categoryName.toLowerCase();
         // Appeler la méthode correspondante dans le Repository pour vérifier l'existence de la catégorie
-        return categoryRepository.doesCategoryExist(categoryName);
+        return categoryRepository.doesCategoryExist(normalizedCategoryName);
     }
 
     @Override
@@ -51,9 +55,15 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
-    public void deleteCategory(int categoryId) throws SQLException {
+    public void deleteCategory(int categoryId) throws SQLException, CategoryNotFoundException {
+        // Vérifier si la catégorie existe avant de tenter de la supprimer
+        if (!categoryRepository.doesCategoryExistById(categoryId)) {
+            throw new CategoryNotFoundException("La catégorie avec l'ID " + categoryId + " n'a pas été trouvée.");
+        }
         categoryRepository.deleteCategory(categoryId);
     }
+
+
 
     @Override
     public List<Category> findAllCategories() throws SQLException {
